Move header components to header/ folder
