+++ next/components/layout.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import clsx from "clsx";
 
 import { Footer } from "@/components/footer";
-import { Header } from "@/components/header";
+import { Header } from "@/components/header/header";
 import {
   PreviewBanner,
   useIsPreviewBannerVisible,
